backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic-settings==2.7.0
python-multipart==0.0.20
email-validator==2.2.0
google-auth==2.37.0
requests==2.32.3
